packages/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.8.2
sqlalchemy==2.0.36
python-dotenv==1.0.1
google-generativeai==0.3.2
pytest==8.3.4
pytest-asyncio==0.25.2
httpx==0.28.1
